+++ src/trim_slice.rs
@@ -8,6 +8,8 @@ use alloc::{
 	vec::Vec,
 };
 
+use crate::not_whitespace;
+
 
 
 /// # Trim Slice.
@@ -17,13 +19,14 @@ use alloc::{
 /// `String`/`&str`.
 ///
 /// The trait methods included are:
+///
 /// | Method | Description |
 /// | ------ | ----------- |
 /// | `trim` | Trim leading and trailing (ASCII) whitespace. |
 /// | `trim_start` | Trim leading (ASCII) whitespace. |
 /// | `trim_end` | Trim trailing (ASCII) whitespace. |
 ///
-/// **Note:** because these methods work with individual bytes — rather than chars
+/// **Note:** Because these methods work with individual bytes — rather than chars
 /// — these methods only trim [`u8::is_ascii_whitespace`], not [`char::is_whitespace`].
 pub trait TrimSlice {
 	/// # Trim.
@@ -78,6 +81,7 @@ pub trait TrimSlice {
 /// enjoyed by `String`/`&str`.
 ///
 /// The trait methods included are:
+///
 /// | Method | Description |
 /// | ------ | ----------- |
 /// | `trim_matches` | Trim arbitrary leading and trailing bytes via callback. |
@@ -128,95 +132,20 @@ pub trait TrimSliceMatches {
 }
 
 
-
-impl TrimSlice for &[u8] {
-	/// # Trim.
-	///
-	/// Trim leading and trailing (ASCII) whitespace from a slice.
-	fn trim(&self) -> &[u8] {
-		let start: usize = self.iter()
-			.position(|b| ! b.is_ascii_whitespace())
-			.unwrap_or(0);
-
-		self.iter()
-			.rposition(|b| ! b.is_ascii_whitespace())
-			.map_or_else(|| &self[start..], |p| &self[start..=p])
-	}
-
-	/// # Trim Start.
-	///
-	/// Trim leading (ASCII) whitespace from a slice.
-	fn trim_start(&self) -> &[u8] {
-		self.iter()
-			.position(|b| ! b.is_ascii_whitespace())
-			.map_or(self, |p| &self[p..])
-	}
-
-	/// # Trim End.
-	///
-	/// Trim trailing (ASCII) whitespace from a slice.
-	fn trim_end(&self) -> &[u8] {
-		self.iter()
-			.rposition(|b| ! b.is_ascii_whitespace())
-			.map_or(self, |p| &self[..=p])
-	}
-}
-
-impl TrimSliceMatches for &[u8] {
-	/// # Trim Matches.
-	///
-	/// Trim arbitrary leading and trailing bytes as determined by the provided
-	/// callback, where a return value of `true` means trim.
-	fn trim_matches<F>(&self, cb: F) -> &[u8]
-	where F: Fn(u8) -> bool {
-		let cb = |b: &u8| ! cb(*b);
-
-		let start: usize = self.iter()
-			.position(cb)
-			.unwrap_or(0);
-
-		self.iter()
-			.rposition(cb)
-			.map_or_else(|| &self[start..], |p| &self[start..=p])
-	}
-
-	/// # Trim Start Matches.
-	///
-	/// Trim arbitrary leading bytes as determined by the provided callback,
-	/// where a return value of `true` means trim.
-	fn trim_start_matches<F>(&self, cb: F) -> &[u8]
-	where F: Fn(u8) -> bool {
-		self.iter()
-			.position(|b: &u8| ! cb(*b))
-			.map_or(self, |p| &self[p..])
-	}
-
-	/// # Trim Start Matches.
-	///
-	/// Trim arbitrary leading bytes as determined by the provided callback,
-	/// where a return value of `true` means trim.
-	fn trim_end_matches<F>(&self, cb: F) -> &[u8]
-	where F: Fn(u8) -> bool {
-		self.iter()
-			.rposition(|b: &u8| ! cb(*b))
-			.map_or(self, |p| &self[..=p])
-	}
-}
-
-macro_rules! trim_slice_alloc {
+macro_rules! trim_slice {
 	($($ty:ty),+ $(,)?) => ($(
 		impl TrimSlice for $ty {
 			/// # Trim.
 			///
 			/// Trim leading and trailing (ASCII) whitespace from a slice.
 			fn trim(&self) -> &[u8] {
-				let start: usize = self.iter()
-					.position(|b| ! b.is_ascii_whitespace())
-					.unwrap_or(0);
-
 				self.iter()
-					.rposition(|b| ! b.is_ascii_whitespace())
-					.map_or_else(|| &self[start..], |p| &self[start..=p])
+					.position(not_whitespace)
+					.map_or(&[], |start| {
+						// We know there is an end because there's a beginning.
+						let end = self.iter().rposition(not_whitespace).unwrap();
+						&self[start..=end]
+					})
 			}
 
 			/// # Trim Start.
@@ -224,8 +153,8 @@ macro_rules! trim_slice_alloc {
 			/// Trim leading (ASCII) whitespace from a slice.
 			fn trim_start(&self) -> &[u8] {
 				self.iter()
-					.position(|b| ! b.is_ascii_whitespace())
-					.map_or(&self, |p| &self[p..])
+					.position(not_whitespace)
+					.map_or(&[], |p| &self[p..])
 			}
 
 			/// # Trim End.
@@ -233,8 +162,8 @@ macro_rules! trim_slice_alloc {
 			/// Trim trailing (ASCII) whitespace from a slice.
 			fn trim_end(&self) -> &[u8] {
 				self.iter()
-					.rposition(|b| ! b.is_ascii_whitespace())
-					.map_or(&self, |p| &self[..=p])
+					.rposition(not_whitespace)
+					.map_or(&[], |p| &self[..=p])
 			}
 		}
 
@@ -247,13 +176,13 @@ macro_rules! trim_slice_alloc {
 			where F: Fn(u8) -> bool {
 				let cb = |b: &u8| ! cb(*b);
 
-				let start: usize = self.iter()
-					.position(cb)
-					.unwrap_or(0);
-
 				self.iter()
-					.rposition(cb)
-					.map_or_else(|| &self[start..], |p| &self[start..=p])
+					.position(cb)
+					.map_or(&[], |start| {
+						// We know there is an end because there's a beginning.
+						let end = self.iter().rposition(cb).unwrap();
+						&self[start..=end]
+					})
 			}
 
 			/// # Trim Start Matches.
@@ -264,7 +193,7 @@ macro_rules! trim_slice_alloc {
 			where F: Fn(u8) -> bool {
 				self.iter()
 					.position(|b: &u8| ! cb(*b))
-					.map_or(&self, |p| &self[p..])
+					.map_or(&[], |p| &self[p..])
 			}
 
 			/// # Trim Start Matches.
@@ -275,13 +204,13 @@ macro_rules! trim_slice_alloc {
 			where F: Fn(u8) -> bool {
 				self.iter()
 					.rposition(|b: &u8| ! cb(*b))
-					.map_or(&self, |p| &self[..=p])
+					.map_or(&[], |p| &self[..=p])
 			}
 		}
 	)+);
 }
 
-trim_slice_alloc!(Box<[u8]>, Vec<u8>);
+trim_slice!(&[u8], Box<[u8]>, Vec<u8>);
 
 
 
@@ -293,45 +222,73 @@ mod tests {
 	const T_EMPTY: &[u8] = b"";
 	const T_HELLO: &[u8] = b"hello";
 	const T_HELLO_E: &[u8] = b"hello\t";
-	const T_HELLO_S: &[u8] = b"\thello";
-	const T_HELLO_SE: &[u8] = b"\n  hello \t";
 
 	#[test]
 	fn t_trim() {
-		let tests: &[(&[u8], &[u8])] = &[
-			(T_EMPTY, T_EMPTY),
-			(T_HELLO, T_HELLO),
-			(T_HELLO_S, T_HELLO),
-			(T_HELLO_E, T_HELLO),
-			(T_HELLO_SE, T_HELLO),
+		let tests: [(&str, &str); 6] = [
+			("", ""),
+			(" \t\n\r", ""),
+			("hello", "hello"),
+			("hello\t", "hello"),
+			("\thello", "hello"),
+			("\n  hello world! \t", "hello world!"),
 		];
 
-		for &(src, expected) in tests.iter() {
-			assert_eq!(src.trim(), expected);
-			assert_eq!(Box::<[u8]>::from(src).trim(), expected);
-			assert_eq!(src.to_vec().trim(), expected);
+		for (raw, expected) in tests.iter() {
+			let a = raw.as_bytes();
+			let b = expected.as_bytes();
+			assert_eq!(a.trim(), b);
+
+			let a = a.to_vec();
+			assert_eq!(a.trim(), b);
+
+			let a = a.into_boxed_slice();
+			assert_eq!(a.trim(), b);
 		}
 
+		assert_eq!(T_EMPTY.trim_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!(T_EMPTY.to_vec().trim_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!(Box::<[u8]>::from(T_EMPTY).trim_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
+
+		assert_eq!("  ".as_bytes().trim_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!("  ".as_bytes().to_vec().trim_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!(Box::<[u8]>::from("  ".as_bytes()).trim_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
+
 		assert_eq!(T_HELLO_E.trim_matches(|b| b'h' == b), b"ello\t");
-		assert_eq!(Box::<[u8]>::from(T_HELLO_E).trim_matches(|b| b'h' == b), b"ello\t");
 		assert_eq!(T_HELLO_E.to_vec().trim_matches(|b| b'h' == b), b"ello\t");
+		assert_eq!(Box::<[u8]>::from(T_HELLO_E).trim_matches(|b| b'h' == b), b"ello\t");
 	}
 
 	#[test]
 	fn t_trim_start() {
-		let tests: &[(&[u8], &[u8])] = &[
-			(T_EMPTY, T_EMPTY),
-			(T_HELLO, T_HELLO),
-			(T_HELLO_S, T_HELLO),
-			(T_HELLO_E, T_HELLO_E),
-			(T_HELLO_SE, b"hello \t"),
+		let tests: [(&str, &str); 6] = [
+			("", ""),
+			(" \t\n\r", ""),
+			("hello", "hello"),
+			("hello\t", "hello\t"),
+			("\thello", "hello"),
+			("\n  hello world! \t", "hello world! \t"),
 		];
 
-		for &(src, expected) in tests.iter() {
-			assert_eq!(src.trim_start(), expected);
-			assert_eq!(Box::<[u8]>::from(src).trim_start(), expected);
-			assert_eq!(src.to_vec().trim_start(), expected);
+		for (raw, expected) in tests.iter() {
+			let a = raw.as_bytes();
+			let b = expected.as_bytes();
+			assert_eq!(a.trim_start(), b);
+
+			let a = a.to_vec();
+			assert_eq!(a.trim_start(), b);
+
+			let a = a.into_boxed_slice();
+			assert_eq!(a.trim_start(), b);
 		}
+
+		assert_eq!(T_EMPTY.trim_start_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!(T_EMPTY.to_vec().trim_start_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!(Box::<[u8]>::from(T_EMPTY).trim_start_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
+
+		assert_eq!("  ".as_bytes().trim_start_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!("  ".as_bytes().to_vec().trim_start_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!(Box::<[u8]>::from("  ".as_bytes()).trim_start_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
 
 		assert_eq!(T_HELLO_E.trim_start_matches(|b| b'h' == b), b"ello\t");
 		assert_eq!(Box::<[u8]>::from(T_HELLO_E).trim_start_matches(|b| b'h' == b), b"ello\t");
@@ -340,19 +297,34 @@ mod tests {
 
 	#[test]
 	fn t_trim_end() {
-		let tests: &[(&[u8], &[u8])] = &[
-			(T_EMPTY, T_EMPTY),
-			(T_HELLO, T_HELLO),
-			(T_HELLO_S, T_HELLO_S),
-			(T_HELLO_E, T_HELLO),
-			(T_HELLO_SE, b"\n  hello"),
+		let tests: [(&str, &str); 6] = [
+			("", ""),
+			(" \t\n\r", ""),
+			("hello", "hello"),
+			("hello\t", "hello"),
+			("\thello", "\thello"),
+			("\n  hello world! \t", "\n  hello world!"),
 		];
 
-		for &(src, expected) in tests.iter() {
-			assert_eq!(src.trim_end(), expected);
-			assert_eq!(Box::<[u8]>::from(src).trim_end(), expected);
-			assert_eq!(src.to_vec().trim_end(), expected);
+		for (raw, expected) in tests.iter() {
+			let a = raw.as_bytes();
+			let b = expected.as_bytes();
+			assert_eq!(a.trim_end(), b);
+
+			let a = a.to_vec();
+			assert_eq!(a.trim_end(), b);
+
+			let a = a.into_boxed_slice();
+			assert_eq!(a.trim_end(), b);
 		}
+
+		assert_eq!(T_EMPTY.trim_end_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!(T_EMPTY.to_vec().trim_end_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!(Box::<[u8]>::from(T_EMPTY).trim_end_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
+
+		assert_eq!("  ".as_bytes().trim_end_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!("  ".as_bytes().to_vec().trim_end_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
+		assert_eq!(Box::<[u8]>::from("  ".as_bytes()).trim_end_matches(|b| b.is_ascii_whitespace()), T_EMPTY);
 
 		assert_eq!(T_HELLO_E.trim_matches(|b| b'\t' == b), T_HELLO);
 		assert_eq!(Box::<[u8]>::from(T_HELLO_E).trim_matches(|b| b'\t' == b), T_HELLO);