+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [0.1.1](https://github.com/Blobfolio/trimothy/releases/tag/v0.1.1) - 2022-04-11
+
+## Changes
+
+* Minor performance improvements
+
+## Fixed
+
+* Return empty slice when all bytes match trim predicate.
+* Markdown (docs) formatting issues.
+
+
+
 ## [0.1.0](https://github.com/Blobfolio/trimothy/releases/tag/v0.1.0) - 2022-04-11
 
 Initial release!